+++ .changeset/description-decorator.md
@@ -2,4 +2,4 @@
 "@inversifyjs/http-open-api": minor
 ---
 
-Added `Description` decorator
+Added `OasDescription` decorator
