+++ .changeset/empty-turtles-divide.md
@@ -2,4 +2,4 @@
 "@inversifyjs/http-open-api": minor
 ---
 
-Added `Response` decorator
+Added `OasResponse` decorator
