+++ .changeset/every-chefs-decide.md
@@ -2,4 +2,4 @@
 "@inversifyjs/http-open-api": minor
 ---
 
-Added `Schema` decorator
+Added `OasSchema` decorator
