+++ .changeset/giant-grapes-worry.md
@@ -2,4 +2,4 @@
 "@inversifyjs/http-open-api": minor
 ---
 
-Added `Server` decorator
+Added `OasServer` decorator
