+++ .changeset/operation-id-decorator.md
@@ -2,4 +2,4 @@
 "@inversifyjs/http-open-api": minor
 ---
 
-Added `OperationId` decorator
+Added `OasOperationId` decorator
