+++ .changeset/petite-ducks-attack.md
@@ -2,4 +2,4 @@
 "@inversifyjs/http-open-api": minor
 ---
 
-Added `SchemaProperty` decorator
+Added `OasSchemaProperty` decorator
