+++ .changeset/tasty-feet-sin.md
@@ -2,4 +2,4 @@
 "@inversifyjs/http-open-api": minor
 ---
 
-Added `Summary`
+Added `OasSummary`
