+++ packages/framework/http/libraries/open-api/src/index.ts
@@ -1,14 +1,14 @@
-export { Deprecated } from './metadata/decorators/Deprecated';
-export { Description } from './metadata/decorators/Description';
-export { ExternalDocs } from './metadata/decorators/ExternalDocs';
-export { OperationId } from './metadata/decorators/OperationId';
-export { Response } from './metadata/decorators/Response';
-export { Schema } from './metadata/decorators/Schema';
-export { SchemaProperty } from './metadata/decorators/SchemaProperty';
-export { Security } from './metadata/decorators/Security';
-export { Server } from './metadata/decorators/Server';
-export { Summary } from './metadata/decorators/Summary';
-export { Tag } from './metadata/decorators/Tag';
+export { OasDeprecated } from './metadata/decorators/OasDeprecated';
+export { Description } from './metadata/decorators/OasDescription';
+export { OasExternalDocs } from './metadata/decorators/OasExternalDocs';
+export { OasOperationId } from './metadata/decorators/OasOperationId';
+export { OasResponse } from './metadata/decorators/OasResponse';
+export { OasSchema } from './metadata/decorators/OasSchema';
+export { OasSchemaProperty } from './metadata/decorators/OasSchemaProperty';
+export { OasSecurity } from './metadata/decorators/OasSecurity';
+export { OasServer } from './metadata/decorators/OasServer';
+export { OasSummary } from './metadata/decorators/OasSummary';
+export { OasTag } from './metadata/decorators/OasTag';
 export { SwaggerUiExpressProvider } from './openApi/services/SwaggerUiExpressProvider';
 export { SwaggerUiExpress4Provider } from './openApi/services/SwaggerUiExpress4Provider';
 export { SwaggerUiFastifyProvider } from './openApi/services/SwaggerUiFastifyProvider';
