+++ packages/framework/http/libraries/open-api/src/metadata/decorators/OasDeprecated.ts
@@ -5,7 +5,7 @@ import { updateControllerOpenApiMetadataOperationProperty } from '../actions/upd
 import { buildDefaultControllerOpenApiMetadata } from '../calculations/buildDefaultControllerOpenApiMetadata';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
-export function Deprecated(): MethodDecorator {
+export function OasDeprecated(): MethodDecorator {
   return (target: object, key: string | symbol): void => {
     updateOwnReflectMetadata(
       target.constructor,
