+++ packages/framework/http/libraries/open-api/src/metadata/decorators/OasExternalDocs.ts
@@ -1,23 +1,24 @@
-import { OpenApi3Dot1SecurityRequirementObject } from '@inversifyjs/open-api-types/v3Dot1';
+import { OpenApi3Dot1ExternalDocumentationObject } from '@inversifyjs/open-api-types/v3Dot1';
 import { updateOwnReflectMetadata } from '@inversifyjs/reflect-metadata-utils';
 
 import { controllerOpenApiMetadataReflectKey } from '../../reflectMetadata/data/controllerOpenApiMetadataReflectKey';
-import { updateControllerOpenApiMetadataOperationArrayProperty } from '../actions/updateControllerOpenApiMetadataOperationArrayProperty';
+import { updateControllerOpenApiMetadataOperationProperty } from '../actions/updateControllerOpenApiMetadataOperationProperty';
 import { buildDefaultControllerOpenApiMetadata } from '../calculations/buildDefaultControllerOpenApiMetadata';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
-export function Security(
-  content: OpenApi3Dot1SecurityRequirementObject,
+export function OasExternalDocs(
+  content: OpenApi3Dot1ExternalDocumentationObject,
 ): MethodDecorator {
   return (target: object, key: string | symbol): void => {
     updateOwnReflectMetadata(
       target.constructor,
       controllerOpenApiMetadataReflectKey,
       buildDefaultControllerOpenApiMetadata,
-      updateControllerOpenApiMetadataOperationArrayProperty(
+      updateControllerOpenApiMetadataOperationProperty(
         content,
+        target.constructor,
         key,
-        'security',
+        'externalDocs',
       ),
     );
   };
