+++ packages/framework/http/libraries/open-api/src/metadata/decorators/OasServer.ts
@@ -1,12 +1,15 @@
+import { OpenApi3Dot1ServerObject } from '@inversifyjs/open-api-types/v3Dot1';
 import { updateOwnReflectMetadata } from '@inversifyjs/reflect-metadata-utils';
 
 import { controllerOpenApiMetadataReflectKey } from '../../reflectMetadata/data/controllerOpenApiMetadataReflectKey';
-import { updateControllerOpenApiMetadataSummary } from '../actions/updateControllerOpenApiMetadataSummary';
+import { updateControllerOpenApiMetadataServer } from '../actions/updateControllerOpenApiMetadataServer';
 import { buildDefaultControllerOpenApiMetadata } from '../calculations/buildDefaultControllerOpenApiMetadata';
 import { ControllerOpenApiMetadata } from '../models/ControllerOpenApiMetadata';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
-export function Summary(summary: string): ClassDecorator & MethodDecorator {
+export function OasServer(
+  server: OpenApi3Dot1ServerObject,
+): ClassDecorator & MethodDecorator {
   return (target: object, key?: string | symbol): void => {
     // eslint-disable-next-line @typescript-eslint/no-unsafe-function-type
     const typeTarget: Function =
@@ -16,7 +19,7 @@ export function Summary(summary: string): ClassDecorator & MethodDecorator {
       typeTarget,
       controllerOpenApiMetadataReflectKey,
       buildDefaultControllerOpenApiMetadata,
-      updateControllerOpenApiMetadataSummary(summary, typeTarget, key),
+      updateControllerOpenApiMetadataServer(server, key),
     );
   };
 }
