+++ packages/framework/http/libraries/open-api/src/metadata/decorators/OasSummary.ts
@@ -1,15 +1,12 @@
-import { OpenApi3Dot1ServerObject } from '@inversifyjs/open-api-types/v3Dot1';
 import { updateOwnReflectMetadata } from '@inversifyjs/reflect-metadata-utils';
 
 import { controllerOpenApiMetadataReflectKey } from '../../reflectMetadata/data/controllerOpenApiMetadataReflectKey';
-import { updateControllerOpenApiMetadataServer } from '../actions/updateControllerOpenApiMetadataServer';
+import { updateControllerOpenApiMetadataSummary } from '../actions/updateControllerOpenApiMetadataSummary';
 import { buildDefaultControllerOpenApiMetadata } from '../calculations/buildDefaultControllerOpenApiMetadata';
 import { ControllerOpenApiMetadata } from '../models/ControllerOpenApiMetadata';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
-export function Server(
-  server: OpenApi3Dot1ServerObject,
-): ClassDecorator & MethodDecorator {
+export function OasSummary(summary: string): ClassDecorator & MethodDecorator {
   return (target: object, key?: string | symbol): void => {
     // eslint-disable-next-line @typescript-eslint/no-unsafe-function-type
     const typeTarget: Function =
@@ -19,7 +16,7 @@ export function Server(
       typeTarget,
       controllerOpenApiMetadataReflectKey,
       buildDefaultControllerOpenApiMetadata,
-      updateControllerOpenApiMetadataServer(server, key),
+      updateControllerOpenApiMetadataSummary(summary, typeTarget, key),
     );
   };
 }
