+++ packages/framework/http/libraries/open-api/src/metadata/decorators/OasTag.ts
@@ -1,24 +1,20 @@
-import { OpenApi3Dot1ExternalDocumentationObject } from '@inversifyjs/open-api-types/v3Dot1';
 import { updateOwnReflectMetadata } from '@inversifyjs/reflect-metadata-utils';
 
 import { controllerOpenApiMetadataReflectKey } from '../../reflectMetadata/data/controllerOpenApiMetadataReflectKey';
-import { updateControllerOpenApiMetadataOperationProperty } from '../actions/updateControllerOpenApiMetadataOperationProperty';
+import { updateControllerOpenApiMetadataOperationArrayProperty } from '../actions/updateControllerOpenApiMetadataOperationArrayProperty';
 import { buildDefaultControllerOpenApiMetadata } from '../calculations/buildDefaultControllerOpenApiMetadata';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
-export function ExternalDocs(
-  content: OpenApi3Dot1ExternalDocumentationObject,
-): MethodDecorator {
+export function OasTag(content: string): MethodDecorator {
   return (target: object, key: string | symbol): void => {
     updateOwnReflectMetadata(
       target.constructor,
       controllerOpenApiMetadataReflectKey,
       buildDefaultControllerOpenApiMetadata,
-      updateControllerOpenApiMetadataOperationProperty(
+      updateControllerOpenApiMetadataOperationArrayProperty(
         content,
-        target.constructor,
         key,
-        'externalDocs',
+        'tags',
       ),
     );
   };
